Add files via upload
C++ program to the question below:
Q1)A hexagonal prism of side of base 30 mm and axis 75 mm long resting, on its base on HP such that rectangle face is parallel to VP. It is cut by section plane perpendicular to VP and inclined to HP by 30 .The section plane is passing through the end of extreme lateral edge of prism. Draw the development of lateral surface of the cut prism.